requirements.txt
fastapi
uvicorn[standard]
pypdf
pymupdf
pytesseract
pillow
opencv-python-headless
langgraph
langchain
langchain-google-genai
pydantic
python-dotenv 
fastmcp 
langchain-groq